Refactor auth context storage and fix lint errors
