+++ frontend/src/main.tsx
@@ -2,7 +2,8 @@ import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import './index.css'
 import App from './App.tsx'
-import { AuthProvider, readStoredAuth } from './context/AuthContext.tsx'
+import { AuthProvider } from './context/AuthContext.tsx'
+import { readStoredAuth } from './context/authStorage.ts'
 
 const initialAuth = readStoredAuth()
